# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daily Challenge Bingo API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /participants:
    get:
      operationId: listParticipants
      tags: [participants]
      summary: List all participants (admin only)
      responses:
        "200":
          description: List of participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Participant"

    post:
      operationId: createParticipant
      tags: [participants]
      summary: Create a participant (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
  /participants/{id}:
    get:
      operationId: getParticipant
      tags: [participants]
      summary: Get a participant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        "404":
          description: Not found

    patch:
      operationId: updateParticipant
      tags: [participants]
      summary: Update a participant (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"

    delete:
      operationId: deleteParticipant
      tags: [participants]
      summary: Delete a participant (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /participants/{id}/progress:
    get:
      operationId: getParticipantProgress
      tags: [participants]
      summary: Get progress for a participant (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Participant progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantProgress"

  # ── Bingo Cards ───────────────────────────────────────────────────────────────
...
  /reflections:
    get:
      operationId: listReflections
      tags: [reflections]
      summary: List reflections (admin sees all, participant sees own)
      responses:
        "200":
          description: List of reflections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reflection"

    post:
      operationId: createReflection
      tags: [reflections]
      summary: Submit today's reflection (participant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReflectionInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reflection"
  /reflections/{id}:
    get:
      operationId: getReflection
      tags: [reflections]
      summary: Get a single reflection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reflection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reflection"

    patch:
      operationId: updateReflection
      tags: [reflections]
      summary: Update a reflection (participant, same day only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReflectionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reflection"

  # ── Dashboard ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Participant:
      type: object
      required: [id, name, username, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, participant]
        createdAt:
          type: string
          format: date-time
    ParticipantInput:
      type: object
      required: [name, username, password]
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        email:
          type: string
    ParticipantUpdate:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        email:
          type: string
    ParticipantProgress:
      type: object
      required: [participantId, name, totalBoxes, hidden, revealed, completed, reflectionsCount]
      properties:
        participantId:
          type: integer
        name:
          type: string
        totalBoxes:
          type: integer
        hidden:
          type: integer
        revealed:
          type: integer
        completed:
          type: integer
        reflectionsCount:
          type: integer

    # ── Bingo Cards ────────────────────────────────────────────────────────────
...
    Reflection:
      type: object
      required: [id, userId, date, whatIChose, whatIDid, impact, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        participantName:
          type: string
        date:
          type: string
          format: date
        whatIChose:
          type: string
        whatIDid:
          type: string
        impact:
          type: string
        createdAt:
          type: string
          format: date-time
    ReflectionInput:
      type: object
      required: [whatIChose, whatIDid, impact]
      properties:
        whatIChose:
          type: string
        whatIDid:
          type: string
        impact:
          type: string
    ReflectionUpdate:
      type: object
      properties:
        whatIChose:
          type: string
        whatIDid:
          type: string
        impact:
          type: string

    # ── Dashboard ──────────────────────────────────────────────────────────────
...