# lib/api-zod/src/generated/types/participantProgress.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Daily Challenge Bingo API
 * OpenAPI spec version: 0.1.0
 */

export interface ParticipantProgress {
  participantId: number;
  name: string;
  totalBoxes: number;
  hidden: number;
  revealed: number;
  completed: number;
  reflectionsCount: number;
}