# lib/api-zod/src/generated/types/reflection.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Daily Challenge Bingo API
 * OpenAPI spec version: 0.1.0
 */

export interface Reflection {
  id: number;
  userId: number;
  participantName?: string;
  date: Date;
  whatIChose: string;
  whatIDid: string;
  impact: string;
  createdAt: Date;
}