# lib/api-zod/src/generated/types/reflectionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Daily Challenge Bingo API
 * OpenAPI spec version: 0.1.0
 */

export interface ReflectionInput {
  whatIChose: string;
  whatIDid: string;
  impact: string;
}